EFT-Helper-Tarkov-Telegram-Bot/pyproject.toml
[tool.poetry]
name = "eft-helper"
version = "1.0.0"
description = "Профессиональный Telegram-бот для игроков Escape from Tarkov"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
package-mode = false
keywords = ["telegram", "bot", "tarkov", "eft", "escape-from-tarkov"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
aiogram = "^3.15.0"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.0"
rapidfuzz = "^3.0.0"
feedparser = "^6.0.0"
python-dateutil = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
